+++ Example/WTFCatView.swift
@@ -3,7 +3,7 @@
 //  LoadableView
 //
 //  Created by Denys Telezhkin on 05.03.16.
-//  Copyright © 2016 Denys Telezhkin. All rights reserved.
+//  Copyright © 2018 Denys Telezhkin. All rights reserved.
 //
 
 import UIKit